Restructure example folders into mechanical systems (eom) and variation linearization
