+++ output/PointMass2.tex
@@ -1,0 +1,4 @@
+PointMass2: Equations of Motion
+\begin{itemize}
+\item Lag1: $u+e_3 (-1.0) m g+\ddot{x} (-1.0) m = 0$
+\end{itemize}
