+++ src/main/scala/hybridrobotics/dynamics/Main.scala
@@ -1,15 +1,15 @@
 package hybridrobotics.dynamics
 
-import hybridrobotics.dynamics.mechanical_systems._
+import hybridrobotics.dynamics.examples._
+
 
 // TODO create and test generalized quadrotor flexible hose system
 // TODO test the code just using sbt and terminal
 // TODO add test cases for DataTypes
 object Main extends App {
 
-  ExampleSystem.main()
-//  QuadrotorWithLoad.main()
-
+//  ExampleSystem.main()
+  mechanical_systems.PointMass.main()
 
 }
