+++ src/main/scala/hybridrobotics/dynamics/examples/mechanical_systems/QuadrotorWithContinuum.scala
@@ -1,4 +1,4 @@
-package hybridrobotics.dynamics.mechanical_systems
+package hybridrobotics.dynamics.examples.mechanical_systems
 
 import java.io.{File, PrintWriter}
