+++ src/main/scala/hybridrobotics/dynamics/examples/variation_linearization/PointMass.scala
@@ -1,0 +1,5 @@
+package hybridrobotics.dynamics.examples.variation_linearization
+
+object PointMass {
+
+}
